requirements/requirements.txt
Django ==3.1.6
djangorestframework==3.12.2
drf-yasg==1.20.0
Pillow==5.1.0

